High Quality Youtube Playback
The way Youtube delivers media files now is by splitting them up into 2
separate streams of audio only and video only. Unfortunately, it isn't
possible to get a fully muxed stream that has decent quality anymore.

I attempted to mux using ffmpeg, but the speed was nowhere near
sufficient to give a smooth stream. In the end, the only option that
really worked out is have Youtube-dl download the video and audio, mux
it into a local file and then have omxplayer play this file.

The video file gets removed as soon as playback is finished as to not
cause hard drive clutter on already limited space like an RPi.

Closes #60